core_ui_service/requirements.txt
fin123-core==0.7.4
fastapi>=0.110
uvicorn[standard]>=0.27
